app/requirements.txt
pycryptodome==3.20.0
pydantic==2.6.2
fastapi==0.109.2
oracledb==2.1.1
uvicorn==0.27.1
gunicorn==22.0.0